+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Unreleased
 
  * `NftId.[to|from]string()` now uses format `1.2.3/4` instead of `1.2.3@4`
-
+ * `TokenNftInfoQuery.setNftId()`
+ * `TokenNftInfoQuery.byNftId()` - Use `TokenNftInfoQuery.setNftId()` instead
+ * `TokenNftInfoQuery.byAccountId()` with no replacement
+ * `TokenNftInfoQuery.byTokenId()` with no replacement
+ * `TokenNftInfoQuery.[set|get]Start()` with no replacement
+ * `TokenNftInfoQuery.[set|get]End()` with no replacement
 
 ### v2.0.13